Optimize Wind Generator

+++ src/main/java/mekanism/generators/client/render/RenderWindGenerator.java
@@ -42,7 +42,7 @@ public class RenderWindGenerator extends TileEntitySpecialRenderer<TileEntityWin
                   + ((tileEntity.getPos().getY() + 4F) / TileEntityWindGenerator.SPEED_SCALED) * partialTick) % 360;
         }
 
-        model.render(0.0625F, angle);
+        model.render(0.0625F, (float) angle);
         GlStateManager.popMatrix();
     }
 }
